+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 3.0.0-dev
 - BC break: removed support for *.phpt files
+- BC break: removed Assert
 
 Version 2.1.0
 - allowed customization of test suite creation in automated tests runner