+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 3.0.0-dev
 - BC break: removed Assert
 - BC break: removed Environment
 - BC break: removed data and testSuit annotations
+- allowed customization of finding test suites
 
 Version 2.1.0
 - allowed customization of test suite creation in automated tests runner