+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Version 3.0.0-dev
 - BC break: removed Assert
 - BC break: removed Environment
 - BC break: removed data and testSuit annotations
-- allowed customization of finding test suites
+- allowed customization of finding test suites in automated tests runner
 
 Version 2.1.0
 - allowed customization of test suite creation in automated tests runner