+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Version 3.0.0-dev
 - automated tests runner's output can be colorized
 - dropped support for Composer 1
 - raised minimal version of PHP to 8.0
-- removed support for phpDoc annotations
+- BC break: removed support for phpDoc annotations
 
 Version 2.1.0
 - allowed customization of test suite creation in automated tests runner